EE-128: Move mapping methods (impl From) to mappings.rs file.
Rename apply_effect_result_to_ipc to grpc_response_from_commit_result.

Get rid of intermediate variables to make the code clearer.
