+++ execution-engine/comm/src/engine_server/mappings.rs
@@ -1,9 +1,15 @@
 use std::collections::BTreeMap;
 use std::convert::{TryFrom, TryInto};
 
+use execution_engine::engine::{Error as EngineError, ExecutionResult};
+use execution_engine::execution::Error as ExecutionError;
+use ipc;
+use storage::error::{Error::*, RootNotFound};
+use storage::{gs, history, op, transform};
+
 /// Helper method for turning instances of Value into Transform::Write.
-fn transform_write(v: common::value::Value) -> Result<storage::transform::Transform, ParsingError> {
-    Ok(storage::transform::Transform::Write(v))
+fn transform_write(v: common::value::Value) -> Result<transform::Transform, ParsingError> {
+    Ok(transform::Transform::Write(v))
 }
 
 #[derive(Debug)]
@@ -14,11 +20,11 @@ fn parse_error<T>(message: String) -> Result<T, ParsingError> {
     Err(ParsingError(message))
 }
 
-impl TryFrom<&super::ipc::Transform> for storage::transform::Transform {
+impl TryFrom<&super::ipc::Transform> for transform::Transform {
     type Error = ParsingError;
-    fn try_from(tr: &super::ipc::Transform) -> Result<storage::transform::Transform, ParsingError> {
+    fn try_from(tr: &super::ipc::Transform) -> Result<transform::Transform, ParsingError> {
         if tr.has_identity() {
-            Ok(storage::transform::Transform::Identity)
+            Ok(transform::Transform::Identity)
         } else if tr.has_add_keys() {
             let keys_map = tr
                 .get_add_keys()
@@ -29,11 +35,9 @@ impl TryFrom<&super::ipc::Transform> for storage::transform::Transform {
                     local_nk.get_key().try_into().map(|k| (local_nk.name, k))
                 })
                 .collect::<Result<BTreeMap<String, common::key::Key>, ParsingError>>()?;
-            Ok(storage::transform::Transform::AddKeys(keys_map))
+            Ok(transform::Transform::AddKeys(keys_map))
         } else if tr.has_add_i32() {
-            Ok(storage::transform::Transform::AddInt32(
-                tr.get_add_i32().value,
-            ))
+            Ok(transform::Transform::AddInt32(tr.get_add_i32().value))
         } else if tr.has_write() {
             let v = tr.get_write().get_value();
             if v.has_integer() {
@@ -133,33 +137,30 @@ impl From<common::value::Value> for super::ipc::Value {
     }
 }
 
-impl From<storage::transform::Transform> for super::ipc::Transform {
-    fn from(tr: storage::transform::Transform) -> Self {
+impl From<transform::Transform> for super::ipc::Transform {
+    fn from(tr: transform::Transform) -> Self {
         let mut t = super::ipc::Transform::new();
         match tr {
-            storage::transform::Transform::Identity => {
+            transform::Transform::Identity => {
                 t.set_identity(super::ipc::TransformIdentity::new());
             }
-            storage::transform::Transform::Write(v) => {
+            transform::Transform::Write(v) => {
                 let mut tw = super::ipc::TransformWrite::new();
                 tw.set_value(v.into());
                 t.set_write(tw)
             }
-            storage::transform::Transform::AddInt32(i) => {
+            transform::Transform::AddInt32(i) => {
                 let mut add = super::ipc::TransformAddInt32::new();
                 add.set_value(i);
                 t.set_add_i32(add);
             }
-            storage::transform::Transform::AddKeys(keys_map) => {
+            transform::Transform::AddKeys(keys_map) => {
                 let mut add = super::ipc::TransformAddKeys::new();
                 let keys = URefMap(keys_map).into();
                 add.set_value(protobuf::RepeatedField::from_vec(keys));
                 t.set_add_keys(add);
             }
-            storage::transform::Transform::Failure(storage::transform::TypeMismatch {
-                expected,
-                found,
-            }) => {
+            transform::Transform::Failure(transform::TypeMismatch { expected, found }) => {
                 let mut fail = super::ipc::TransformFailure::new();
                 let mut typemismatch_err = super::ipc::StorageTypeMismatch::new();
                 typemismatch_err.set_expected(expected.to_owned());
@@ -254,26 +255,26 @@ impl TryFrom<&super::ipc::Key> for common::key::Key {
     }
 }
 
-impl From<storage::op::Op> for super::ipc::Op {
-    fn from(op: storage::op::Op) -> super::ipc::Op {
+impl From<op::Op> for super::ipc::Op {
+    fn from(op: op::Op) -> super::ipc::Op {
         let mut ipc_op = super::ipc::Op::new();
         match op {
-            storage::op::Op::Read => ipc_op.set_read(super::ipc::ReadOp::new()),
-            storage::op::Op::Write => ipc_op.set_write(super::ipc::WriteOp::new()),
-            storage::op::Op::Add => ipc_op.set_add(super::ipc::AddOp::new()),
-            storage::op::Op::NoOp => ipc_op.set_noop(super::ipc::NoOp::new()),
+            op::Op::Read => ipc_op.set_read(super::ipc::ReadOp::new()),
+            op::Op::Write => ipc_op.set_write(super::ipc::WriteOp::new()),
+            op::Op::Add => ipc_op.set_add(super::ipc::AddOp::new()),
+            op::Op::NoOp => ipc_op.set_noop(super::ipc::NoOp::new()),
         };
         ipc_op
     }
 }
 
 /// Transforms gRPC TransformEntry into domain tuple of (Key, Transform).
-impl TryFrom<&super::ipc::TransformEntry> for (common::key::Key, storage::transform::Transform) {
+impl TryFrom<&super::ipc::TransformEntry> for (common::key::Key, transform::Transform) {
     type Error = ParsingError;
     fn try_from(from: &super::ipc::TransformEntry) -> Result<Self, ParsingError> {
         if from.has_key() {
             if from.has_transform() {
-                let t: storage::transform::Transform = from.get_transform().try_into()?;
+                let t: transform::Transform = from.get_transform().try_into()?;
                 let key = from.get_key().try_into()?;
                 Ok((key, t))
             } else {
@@ -285,8 +286,8 @@ impl TryFrom<&super::ipc::TransformEntry> for (common::key::Key, storage::transf
     }
 }
 
-impl From<storage::gs::ExecutionEffect> for super::ipc::ExecutionEffect {
-    fn from(ee: storage::gs::ExecutionEffect) -> super::ipc::ExecutionEffect {
+impl From<gs::ExecutionEffect> for super::ipc::ExecutionEffect {
+    fn from(ee: gs::ExecutionEffect) -> super::ipc::ExecutionEffect {
         let mut eff = super::ipc::ExecutionEffect::new();
         let ipc_ops: Vec<super::ipc::OpEntry> =
             ee.0.iter()
@@ -313,5 +314,238 @@ impl From<storage::gs::ExecutionEffect> for super::ipc::ExecutionEffect {
         eff.set_op_map(protobuf::RepeatedField::from_vec(ipc_ops));
         eff.set_transform_map(protobuf::RepeatedField::from_vec(ipc_tran));
         eff
+    }
+}
+
+impl From<RootNotFound> for ipc::RootNotFound {
+    fn from(err: RootNotFound) -> ipc::RootNotFound {
+        let RootNotFound(missing_root_hash) = err;
+        let mut root_missing_err = ipc::RootNotFound::new();
+        root_missing_err.set_hash(missing_root_hash.to_vec());
+        root_missing_err
+    }
+}
+
+impl From<ExecutionResult> for ipc::DeployResult {
+    fn from(er: ExecutionResult) -> ipc::DeployResult {
+        match er {
+            ExecutionResult {
+                result: Ok(effects),
+                cost,
+            } => {
+                let mut ipc_ee = effects.into();
+                let mut deploy_result = ipc::DeployResult::new();
+                deploy_result.set_effects(ipc_ee);
+                deploy_result.set_cost(cost);
+                deploy_result
+            }
+            ExecutionResult {
+                result: Err(err),
+                cost,
+            } => {
+                match err {
+                    // TODO(mateusz.gorski): Fix error model for the storage errors.
+                    // We don't have separate IPC messages for storage errors
+                    // so for the time being they are all reported as "wasm errors".
+                    EngineError::StorageError(storage_err) => {
+                        let mut err = match storage_err {
+                            KeyNotFound(key) => {
+                                let msg = format!("Key {:?} not found.", key);
+                                wasm_error(msg)
+                            }
+                            RkvError(error_msg) => wasm_error(error_msg),
+                            TransformTypeMismatch(transform::TypeMismatch { expected, found }) => {
+                                let msg = format!(
+                                    "Type mismatch. Expected {:?}, found {:?}",
+                                    expected, found
+                                );
+                                wasm_error(msg)
+                            }
+                            BytesRepr(bytesrepr_err) => {
+                                let msg =
+                                    format!("Error with byte representation: {:?}", bytesrepr_err);
+                                wasm_error(msg)
+                            }
+                        };
+                        err.set_cost(cost);
+                        err
+                    }
+                    EngineError::PreprocessingError(err_msg) => {
+                        let mut err = wasm_error(err_msg);
+                        err.set_cost(cost);
+                        err
+                    }
+                    EngineError::ExecError(exec_error) => match exec_error {
+                        ExecutionError::GasLimit => {
+                            let mut deploy_result = ipc::DeployResult::new();
+                            let mut deploy_error = ipc::DeployError::new();
+                            deploy_error.set_gasErr(ipc::OutOfGasError::new());
+                            deploy_result.set_error(deploy_error);
+                            deploy_result.set_cost(cost);
+                            deploy_result
+                        }
+                        // TODO(mateusz.gorski): Be more specific about execution errors
+                        other => {
+                            let msg = format!("{:?}", other);
+                            let mut err = wasm_error(msg);
+                            err.set_cost(cost);
+                            err
+                        }
+                    },
+                }
+            }
+        }
+    }
+}
+
+pub fn grpc_response_from_commit_result<R, H>(
+    input: Result<storage::history::CommitResult, H::Error>,
+) -> ipc::CommitResponse
+where
+    R: gs::DbReader,
+    H: history::History<R>,
+    H::Error: Into<ipc::RootNotFound>,
+{
+    match input {
+        Err(err) => {
+            let mut ipc_err = err.into();
+            let mut tmp_res = ipc::CommitResponse::new();
+            tmp_res.set_missing_prestate(ipc_err);
+            tmp_res
+        }
+        Ok(history::CommitResult::Success(post_state_hash)) => {
+            println!("Effects applied. New state hash is: {:?}", post_state_hash);
+            let mut commit_result = ipc::CommitResult::new();
+            let mut tmp_res = ipc::CommitResponse::new();
+            commit_result.set_poststate_hash(post_state_hash.to_vec());
+            tmp_res.set_success(commit_result);
+            tmp_res
+        }
+        // TODO(mateusz.gorski): We should be more specific about errors here.
+        Ok(history::CommitResult::Failure(storage_error)) => {
+            println!("Error {:?} when applying effects", storage_error);
+            let mut err = ipc::PostEffectsError::new();
+            let mut tmp_res = ipc::CommitResponse::new();
+            err.set_message(format!("{:?}", storage_error));
+            tmp_res.set_failed_transform(err);
+            tmp_res
+        }
+    }
+}
+
+fn wasm_error(msg: String) -> ipc::DeployResult {
+    let mut deploy_result = ipc::DeployResult::new();
+    let mut deploy_error = ipc::DeployError::new();
+    let mut err = ipc::WasmError::new();
+    err.set_message(msg.to_owned());
+    deploy_error.set_wasmErr(err);
+    deploy_result.set_error(deploy_error);
+    deploy_result
+}
+
+#[cfg(test)]
+mod tests {
+    use super::wasm_error;
+    use common::key::Key;
+    use execution_engine::engine::{Error as EngineError, ExecutionResult};
+    use shared::newtypes::Blake2bHash;
+    use std::collections::HashMap;
+    use std::convert::TryInto;
+    use storage::gs::ExecutionEffect;
+    use storage::transform::Transform;
+
+    // Test that wasm_error function actually returns DeployResult with result set to WasmError
+    #[test]
+    fn wasm_error_result() {
+        let error_msg = "WasmError";
+        let mut result = wasm_error(error_msg.to_owned());
+        assert!(result.has_error());
+        let mut ipc_error = result.take_error();
+        assert!(ipc_error.has_wasmErr());
+        let ipc_wasm_error = ipc_error.take_wasmErr();
+        let ipc_error_msg = ipc_wasm_error.get_message();
+        assert_eq!(ipc_error_msg, error_msg);
+    }
+
+    #[test]
+    fn deploy_result_to_ipc_missing_root() {
+        let root_hash: Blake2bHash = [1u8; 32].into();
+        let mut result: super::ipc::RootNotFound = storage::error::RootNotFound(root_hash).into();
+        let ipc_missing_hash = result.take_hash();
+        assert_eq!(root_hash.to_vec(), ipc_missing_hash);
+    }
+
+    #[test]
+    fn deploy_result_to_ipc_success() {
+        let input_transforms: HashMap<Key, Transform> = {
+            let mut tmp_map = HashMap::new();
+            tmp_map.insert(Key::Account([1u8; 20]), Transform::AddInt32(10));
+            tmp_map
+        };
+        let execution_effect: ExecutionEffect =
+            ExecutionEffect(HashMap::new(), input_transforms.clone());
+        let cost: u64 = 123;
+        let execution_result: ExecutionResult = ExecutionResult::success(execution_effect, cost);
+        let mut ipc_deploy_result: super::ipc::DeployResult = execution_result.into();
+        assert_eq!(ipc_deploy_result.get_cost(), cost);
+
+        // Extract transform map from the IPC message and parse it back to the domain
+        let ipc_transforms: HashMap<Key, Transform> = {
+            let mut ipc_effects = ipc_deploy_result.take_effects();
+            let ipc_effects_tnfs = ipc_effects.take_transform_map().into_vec();
+            ipc_effects_tnfs
+                .iter()
+                .map(|e| e.try_into())
+                .collect::<Result<HashMap<Key, Transform>, _>>()
+                .unwrap()
+        };
+        assert_eq!(&input_transforms, &ipc_transforms);
+    }
+
+    fn into_execution_failure<E: Into<EngineError>>(error: E, cost: u64) -> ExecutionResult {
+        ExecutionResult::failure(error.into(), cost)
+    }
+
+    fn test_cost<E: Into<EngineError>>(expected_cost: u64, err: E) -> u64 {
+        let execution_failure = into_execution_failure(err, expected_cost);
+        let ipc_deploy_result: super::ipc::DeployResult = execution_failure.into();
+        ipc_deploy_result.get_cost()
+    }
+
+    #[test]
+    fn storage_error_has_cost() {
+        use storage::error::Error::*;
+        let cost: u64 = 100;
+        assert_eq!(test_cost(cost, KeyNotFound(Key::Account([1u8; 20]))), cost);
+        assert_eq!(test_cost(cost, RkvError("Error".to_owned())), cost);
+        let type_mismatch = storage::transform::TypeMismatch {
+            expected: "expected".to_owned(),
+            found: "found".to_owned(),
+        };
+        assert_eq!(test_cost(cost, TransformTypeMismatch(type_mismatch)), cost);
+        let bytesrepr_err = common::bytesrepr::Error::EarlyEndOfStream;
+        assert_eq!(test_cost(cost, BytesRepr(bytesrepr_err)), cost);
+    }
+
+    #[test]
+    fn preprocessing_err_has_cost() {
+        let cost: u64 = 100;
+        // it doesn't matter what error type it is
+        let preprocessing_error = wasm_prep::PreprocessingError::NoExportSection;
+        assert_eq!(test_cost(cost, preprocessing_error), cost);
+    }
+
+    #[test]
+    fn exec_err_has_cost() {
+        let cost: u64 = 100;
+        // GasLimit error is treated differently at the moment so test separately
+        assert_eq!(
+            test_cost(cost, execution_engine::execution::Error::GasLimit),
+            cost
+        );
+        // for the time being all other execution errors are treated in the same way
+        let forged_ref_error =
+            execution_engine::execution::Error::ForgedReference(Key::Account([1u8; 20]));
+        assert_eq!(test_cost(cost, forged_ref_error), cost);
     }
 }
